+++ tests/test_docstrings/test_numpy.py
@@ -97,13 +97,8 @@ def test_empty_indented_lines_in_section_with_items(parse_numpy: ParserType) -> 
     """
     docstring = "Returns\n-------\nonly_item : type\n    Description.\n    \n    \n\nSomething."
     sections, _ = parse_numpy(docstring)
-    assert len(sections) == 2
-    assert len(sections[0].value) == 1
-
-    # allow_section_blank_line requires at least 2 newlines to create a new section
-    sections2, _ = parse_numpy(docstring, allow_section_blank_line=True)
-    assert len(sections2) == 1
-    assert len(sections2[0].value) == 2
+    assert len(sections) == 1
+    assert len(sections[0].value) == 2
 
 
 def test_doubly_indented_lines_in_section_items(parse_numpy: ParserType) -> None:
@@ -793,6 +788,23 @@ def test_examples_section_as_last(parse_numpy: ParserType) -> None:
     assert sections[1].kind is DocstringSectionKind.examples
 
 
+def test_blank_lines_in_section(parse_numpy: ParserType) -> None:
+    """Support blank lines in the middle of sections.
+
+    Parameters:
+        parse_numpy: Fixture parser.
+    """
+    docstring = """
+        Examples
+        --------
+        Line 1.
+
+        Line 2.
+    """
+    sections, _ = parse_numpy(docstring)
+    assert len(sections) == 1
+
+
 # =============================================================================================
 # Attributes sections
 def test_retrieve_attributes_annotation_from_parent(parse_numpy: ParserType) -> None:
@@ -943,6 +955,28 @@ def test_detect_optional_flag(parse_numpy: ParserType) -> None:
     assert sections[0].value[1].default == "b''"
     assert sections[0].value[2].annotation == "bytes"
     assert sections[0].value[2].default == "b''"
+
+
+@pytest.mark.parametrize("newlines", [1, 2, 3])
+def test_blank_lines_in_item_descriptions(parse_numpy: ParserType, newlines: int) -> None:
+    """Support blank lines in the middle of item descriptions.
+
+    Parameters:
+        parse_numpy: Fixture parser.
+        newlines: Number of new lines between item summary and its body.
+    """
+    nl = "\n"
+    nlindent = "\n" + " " * 12
+    docstring = f"""
+        Parameters
+        ----------
+        a : str
+            Summary.{nlindent * newlines}Body.
+    """
+    sections, _ = parse_numpy(docstring)
+    assert len(sections) == 1
+    assert sections[0].value[0].annotation == "str"
+    assert sections[0].value[0].description == f"Summary.{nl * newlines}Body."
 
 
 # =============================================================================================